Niclas branch2

+++ src/sample/Controller.java
@@ -1,21 +1,21 @@
 package sample;
-        import javafx.collections.FXCollections;
-        import javafx.scene.control.*;
-        import javafx.scene.control.Label;
-        import javafx.scene.control.ListView;
-        import javafx.scene.control.SelectionMode;
-        import javafx.scene.input.KeyCode;
-        import javafx.scene.input.KeyEvent;
-        import javafx.scene.layout.Pane;
-        import javafx.scene.media.*;
-        import javafx.collections.ObservableList;
-        import javafx.fxml.FXML;
-        import javafx.fxml.Initializable;
-        import java.io.*;
-        import java.net.*;
-        import java.util.ArrayList;
-        import java.util.ResourceBundle;
-        import java.util.StringTokenizer;
+import javafx.collections.FXCollections;
+import javafx.scene.control.*;
+import javafx.scene.control.Label;
+import javafx.scene.control.ListView;
+import javafx.scene.control.SelectionMode;
+import javafx.scene.input.KeyCode;
+import javafx.scene.input.KeyEvent;
+import javafx.scene.layout.Pane;
+import javafx.scene.media.*;
+import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.fxml.Initializable;
+import java.io.*;
+import java.net.*;
+import java.util.ArrayList;
+import java.util.ResourceBundle;
+import java.util.StringTokenizer;
 
 public class Controller implements Initializable {
     @FXML
@@ -34,7 +34,6 @@ public class Controller implements Initializable {
     private ListView<Playlist> lvPlaylist;
     @FXML
     private ListView<Songs> lvSongList;
-    @FXML
     private MediaPlayer mp;
     private Media me;
 
@@ -152,23 +151,29 @@ public class Controller implements Initializable {
         // If autoplay is turned of the method play(), stop(), pause() etc controls how/when medias are played
         mp.setAutoPlay(false);
 
+        //getNextSongID(selectedSong, myPlaylist);
         mp.setOnEndOfMedia(() -> {
             selectNextSong();
             handlePlay();
         });
 
-        //getNextSongID(selectedSong, myPlaylist);
 
         mp.play();
 
     }
 
+    /**
+     *Handler for the pause button
+     */
     @FXML
     private void handlePause() {
         mp.pause();
         isPaused = true;
     }
 
+    /**
+     * Handler for the stop button
+     */
     @FXML
     private void handleStop() {
         mp.stop();
@@ -177,11 +182,11 @@ public class Controller implements Initializable {
 
     @FXML
     /**
-     *
+     * This method handles the continue button. It continues playing the song from where it was paused.
      */
     public void handleContinue() {
         Songs selectedSong = lvSongList.getSelectionModel().getSelectedItem();
-
+        // Continues the song from where it was paused, but only if it has been paused.
         if (!isPaused) {
             String path = new File("src/sample/media/" + selectedSong.getPath()).getAbsolutePath();
             // Create new Media object (the actual media content)
@@ -197,15 +202,18 @@ public class Controller implements Initializable {
 
     /**
      * This method updates the ListView with songs to the current content that needs to be there.
-     *
      * @param selectedPlaylist is used to get a list of all songs currently in the application.
      */
     public void updateSonglist(Playlist selectedPlaylist) {
+        // Sets the items to all the items in the selected playlist.
         items.setAll(selectedPlaylist.songList);
+        // Sets the ListView to the new list of items.
         lvSongList.setItems(items);
     }
 
-    // Creates a pop up window for the user input
+    /**
+     * This methods creates a pop-up window for the user input
+     */
     @FXML
     public void handleCreatePlaylist ()
     {
@@ -213,7 +221,9 @@ public class Controller implements Initializable {
         pAddPlaylist.setOpacity(1.0);
     }
 
-    //Gets user input and creates a playlist with the name specified by the user
+    /**
+     * This method gets the user input and creates a playlist with the name specified by the user
+     */
     @FXML
     public void handleAddPlaylist ()
     {
@@ -245,24 +255,33 @@ public class Controller implements Initializable {
         tfPlaylistName.setText("");
     }
 
-    // Deletes the selected playlist
+    /**
+     * This method deletes the selected playlist from the listview and the database
+     */
     public void handleDeletePlaylist ()
     {
         Playlist selectedPlaylist = lvPlaylist.getSelectionModel().getSelectedItem();
 
-        DB.deleteSQL("Delete from tblPlaylist where fldPlaylistName = '" + selectedPlaylist.getPlaylistName() + "';");
+        if (!selectedPlaylist.getPlaylistName().equals("All Songs")) {
 
 
-        for (int i = 0; i < playlistItems.size(); i++) {
+            DB.deleteSQL("Delete from tblPlaylist where fldPlaylistName = '" + selectedPlaylist.getPlaylistName() + "';");
 
-            playlistsList.remove(playlistItems.get(i));
+
+            for (int i = 0; i < playlistItems.size(); i++) {
+
+                playlistsList.remove(playlistItems.get(i));
+            }
+
+            lvPlaylist.getItems().remove(selectedPlaylist);
+            updatePlaylists(playlistItems);
         }
-
-        lvPlaylist.getItems().remove(selectedPlaylist);
-        updatePlaylists(playlistItems);
     }
 
-    //creates a sequence for a playlist by looking at the songlist of that playlist, and stores it in tblPlaylist
+    /**
+     * Creates a sequence for a playlist by looking at the songlist of that playlist, and stores it in tblPlaylist
+     * @param playlist The playlist for which the sequence should be set
+     */
     public void setPlaylistSequence(Playlist playlist){
         String sequence = "";
         for (int i = 0; i < playlist.getSongList().size(); i++) {
@@ -278,6 +297,11 @@ public class Controller implements Initializable {
         DB.updateSQL("update tblPlaylist set fldSequence = '"+sequence+"' where fldPlaylistName = '"+playlist.getPlaylistName()+"'");
     }
 
+    /**
+     * Gets information of all playlists from DB and creates playlist objects with it.
+     * PlaylistName is stored in list to be used for the creation of the object.
+     * Sequence is split into individual ID's that are used to identify the correct songs to be added to the playlist.
+     */
     public void loadPlaylistsFromDB(){
         ArrayList<String> nameList = new ArrayList<>();
         //Get all name Strings from tblPlaylist
@@ -328,10 +352,13 @@ public class Controller implements Initializable {
     }
 
 
-
+    /**
+     * Moving the listview selection to the next item in the list
+     */
     public void selectNextSong(){
-        //check where in the sequence the selectedSong is
+        //storing the currently selected song
         int songIndex = lvSongList.getSelectionModel().getSelectedIndex();
+        //setting selection to current +1
         songIndex++;
         lvSongList.getSelectionModel().select(songIndex);
     }
@@ -366,7 +393,8 @@ public class Controller implements Initializable {
      */
     private void getPlaylistInformation () {
         // New object that gets the songs(items) from the selected playlist.
-         selectedPlaylist = lvPlaylist.getSelectionModel().getSelectedItem();
+        selectedPlaylist = lvPlaylist.getSelectionModel().getSelectedItem();
+
         // Updates the ListView to show the songs(items) in the selected playlist.
         updateSonglist(selectedPlaylist);
 
@@ -386,15 +414,17 @@ public class Controller implements Initializable {
                 selectedPlaylist.getSongList().remove(i);
             }
 
-    }
+        }
         setPlaylistSequence(selectedPlaylist);
         updateSonglist(selectedPlaylist);
     }
 
 
-    // Creates a pop up window for the user input
 
     @FXML
+    /**
+     * This method creates a pop up window for the user input.
+     */
     public void handleAddToPlaylistPopup() {
         pAddSongToPlaylist.setDisable(false);
         pAddSongToPlaylist.setOpacity(1.0);
@@ -427,19 +457,37 @@ public class Controller implements Initializable {
 
     }
     @FXML
+    /**
+     * This method enables the user to press ENTER/RETURN key on the keyboard to add a song to a playlist.
+     * It's a substitute for pressing the 'OK' button
+     * @param ke is the KeyEvent for the ENTER/Return key.
+     */
     public void handleKeyEnterAddSong (KeyEvent ke) {
+        // An if statement to check if the pressed key is ENTER, if it is it will run the handleSongOK(); method
         if (ke.getCode().equals(KeyCode.ENTER)) {
             handleSongAddOK();
         }
     }
     @FXML
+    /**
+     * This method enables the user to press ENTER/RETURN key on the keyboard add a new playlist with the inputted name.
+     * It's a substitute for pressing the 'OK' button
+     * @param ke is the KeyEvent for the ENTER/Return key.
+     */
     public void handleKeyEnterAddPlaylist (KeyEvent ke) {
+        // An if statement to check if the pressed key is ENTER, if it is it will run the handleAddPlaylist(); method
         if (ke.getCode().equals(KeyCode.ENTER)) {
             handleAddPlaylist();
         }
     }
     @FXML
+    /**
+     * This method enables the user to press ENTER/RETURN key on the keyboard to search for songs.
+     * It's a substitute for pressing the 'search' button
+     * @param ke is the KeyEvent for the ENTER/Return key.
+     */
     public void handleKeyEnterSearch (KeyEvent ke) {
+        // An if statement to check if the pressed key is ENTER, if it is it will run the searchForSongs(); method
         if (ke.getCode().equals(KeyCode.ENTER)) {
             searchForSongs();
         }
